Reduce warnings.  Include <ansidecl.h>; add ATTRIBUTE_UNUSED; change some String's to CString's. Database/database.c Database/nameindex.c Database/recordindex.c Gedcom/lineage.c Gedcom/nodeutls.c Gedcom/readnode.c Gedcom/splitjoin.c Gedcom/writenode.c Interp/builtinlist.c Interp/Includes/builtintable.h Interp/functable.c Interp/interp.c Interp/intrpevent.c Interp/intrpfamily.c Interp/intrpgnode.c Interp/intrpmath.c Interp/intrpseq.c Interp/pnode.c Interp/pvaluetable.c Parser/parse.c     Include <ansidecl.h>. Database/import.c     Include <ansidecl.h>, add ATTRIBUTE_UNUSED Gedcom/Includes/gedcom.h     Add ATTRIBUTE_UNUSED to some variables in FORFAMSS, FORFAMCS, FORTAGVALUES Gedcom/Includes/name.h     Make surname argument of soundex CString, not String Gedcom/gedcom.c     Include <ansidecl.h>; cast strlen result. Gedcom/gnode.c     Include <ansidecl.h>, add cast. Gedcom/name.c     Include <ansidecl.h>; change soundex -- name is now a CString, add casts Interp/builtintable.c     Add prototype for __inset. Interp/Includes/interp.h     Change iistype second arg from int to PNType. Interp/Includes/sequence.h     Add ATTRIBUTE_UNUSED to third argument of FORSEQUENCE Interp/builtin.c     Include <ansidecl.h>; add ATTRIBUTE_UNUSED to some function parameters. Interp/evaluate.c     Include <ansidecl.h>; add ATTRIBUTE_UNUSED; change iistype 2nd arg to PNType Interp/intrpperson.c     Include <ansidecl.h>; add ATTRIBUTE_UNUSED to __firstindi. Interp/pvalue.c     Include <ansidecl.h>; change type in {eq,ne,ge,lt,gt}PValues     from int to PVType. Utils/Includes/date.h     Change first arg of extract_date from String to CString; declare format_date. Utils/date.c     Change arg of set_date_string, extract_date, format_date from String to CString. Interp/sequence.c     Include <ansidecl.h>; remove unused variable from sequenceToGedcom. DataTypes/integertable.c     Use __INT_MAX__ instead of a fake NAN. Utils/Includes/standard.h     Get rid of NAN -- it conflicts with NAN in <math.h>.

+++ Gedcom/gedcom.c
@@ -7,6 +7,8 @@
 //  Last changed on 11 October 2023.
 //
 
+#include <ansidecl.h>		/* ATTRIBUTE_UNUSED */
+
 #include "gedcom.h"
 
 RecordType recordType(GNode *root)
@@ -36,7 +38,7 @@ int compareRecordKeys(Word p, Word q)
     ASSERT(strlen(a) > 1 && strlen(b) > 1);
     if (a[0] != b[0]) return a[0] - b[0];
     if (strlen(a) != strlen(b)) return (int) (strlen(a) - strlen(b));
-    for (int i = 1; i < strlen(a); i++) {
+    for (int i = 1; i < (int)strlen(a); i++) {
         if (a[i] != b[i]) return a[i] - b[i];
     }
     return 0;
